The database
The table relates the chemical shift and H-bond information of residues from 3488 PDB and BMRB IDs that form extended ('E') structures. The table includes backbone and CB the chemical shifts of the residues. information of their direct neighbors and the H-bonds which they are involved into. The H-bonds are identified by DSSP ('H_bond_XX') as well as the search of the proximate oxygens directly in the PDB file.